refactor: 차트를 공통 컴포넌트로 분리
